+++ R/agreement_coefficient.R
@@ -37,7 +37,7 @@
 #' @srrstats {SP4.2} Returns are explicitly documented
 #'
 #' @return
-#' A tibble with columns .metric, .estimator, and .estimate and `nrow(data)` rows of values.
+#' A tibble with columns .metric, .estimator, and .estimate and 1 row of values.
 #' For grouped data frames, the number of rows returned will be the same as the number of groups.
 #' For `_vec()` functions, a single value (or NA).
 #'