+++ R/local_geary.R
@@ -4,6 +4,12 @@
 #' `ww_local_geary_c()` returns the statistic itself, while
 #' `ww_local_geary_pvalue()` returns the associated p value.
 #'
+#' These functions can be used for geographic or projected coordinate reference
+#' systems and expect 2D data.
+#'
+#' @srrstats {SP1.0} Domain of applicability specified above.
+#' @srrstats {SP1.1} Dimensional domain of applicability specified above.
+#'
 #' @srrstats {G1.4} roxygen2 documentation
 #' @srrstats {G2.7} This function relies on yardstick and dplyr and therefore only handles data.frame and vector input.
 #' @srrstats {G2.8} Method dispatch enforces data.frame inputs
@@ -15,13 +21,24 @@
 #' @srrstats {G2.15} Any function may be passed to na_action
 #' @srrstats {G2.16} Any function may be passed to na_action
 #'
+#' @srrstats {SP2.6} Input type requirements are documented.
+#' @srrstats {SP3.0} Users are given total control over weights.
+#' @srrstats {SP3.0a} Users are given total control over weights.
+#' @srrstats {SP3.0b} Users are given total control over weights.
+#' @srrstats {SP3.1} Users are given total control over weights.
 #' @inheritParams ww_global_geary_c
 #' @inheritParams spdep::localC_perm
-#' @param ... Additional arguments passed to [spdep::localC_perm()].
+#' @param ... Additional arguments passed to [spdep::localC()] (for
+#' `ww_local_geary_c()`) or [spdep::localC_perm()] (for
+#' `ww_local_geary_pvalue()`).
+#'
+#' @srrstats {SP4.0} Return values are of a unique format
+#' @srrstats {SP4.0b} Return values are of a unique format
+#' @srrstats {SP4.2} Returns are explicitly documented
 #'
 #' @return
-#' A tibble with columns .metric, .estimator, and .estimate and `nrow(data)` rows of values.
-#' For grouped data frames, the number of rows returned will be the same as the number of groups.
+#' A tibble with columns .metric, .estimator, and .estimate and `nrow(data)`
+#' rows of values.
 #' For `_vec()` functions, a numeric vector of `length(truth)` (or NA).
 #'
 #' @family autocorrelation metrics